fixes #5070 (prototype for Kokkos standard algorithms)

Sample output (`OpenMP`):
```
OMP_PROC_BIND=spread OMP_PLACES=threads ./build/algorithms/unit_tests KokkosAlgorithms_Benchmark_Kokkos_ForEach --benchmark_counters_tabular=true
```

```
Running ./build/algorithms/unit_tests/KokkosAlgorithms_Benchmark_Kokkos_ForEach
Run on (8 X 4700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 12288 KiB (x1)
Load Average: 2.32, 1.61, 1.64
CPU architecture: none
Default Device: N6Kokkos6OpenMPE
GIT_BRANCH: benchmark-algorithms
GIT_CLEAN_STATUS: DIRTY
GIT_COMMIT_DATE: 2023-02-17T13:08:15+01:00
GIT_COMMIT_DESCRIPTION: Add benchmark for Kokkos::for_each
GIT_COMMIT_HASH: 06c406ec2
GPU architecture: none
KOKKOS_COMPILER_CLANG: 1507
KOKKOS_ENABLE_ASM: no
KOKKOS_ENABLE_CXX17: yes
KOKKOS_ENABLE_CXX20: no
KOKKOS_ENABLE_CXX23: no
KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK: yes
KOKKOS_ENABLE_GNU_ATOMICS: no
KOKKOS_ENABLE_HBWSPACE: no
KOKKOS_ENABLE_HWLOC: yes
KOKKOS_ENABLE_INTEL_ATOMICS: no
KOKKOS_ENABLE_INTEL_MM_ALLOC: no
KOKKOS_ENABLE_LIBDL: yes
KOKKOS_ENABLE_LIBRT: no
KOKKOS_ENABLE_OPENMP: yes
KOKKOS_ENABLE_OPENMP_ATOMICS: no
KOKKOS_ENABLE_PRAGMA_IVDEP: no
KOKKOS_ENABLE_PRAGMA_LOOPCOUNT: no
KOKKOS_ENABLE_PRAGMA_UNROLL: no
KOKKOS_ENABLE_PRAGMA_VECTOR: no
KOKKOS_ENABLE_WINDOWS_ATOMICS: no
Kokkos: OpenMP thread_pool_topology[ 1 x 8 x 1 ]
Kokkos Version: 4.0.99
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations Time (Kokkos) Time (std)
-----------------------------------------------------------------------------------------------------------
Algorithms_ForEach<double>/Size:13           0.000 s         0.000 s        368713        1.248u       251n
Algorithms_ForEach<double>/Size:1003         0.000 s         0.000 s         35469        2.877u     12.38u
Algorithms_ForEach<double>/Size:101513       0.002 s         0.001 s           492       30.673u    1.3454m
Algorithms_ForEach<double>/Size:1000003      0.020 s         0.013 s            52      498.657u  0.0197173
```

Sample output (`Serial`):
```
Running ./build/algorithms/unit_tests/KokkosAlgorithms_Benchmark_Kokkos_ForEach
Run on (8 X 4700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 12288 KiB (x1)
Load Average: 2.71, 1.89, 1.71
CPU architecture: none
Default Device: N6Kokkos6SerialE
GIT_BRANCH: benchmark-algorithms
GIT_CLEAN_STATUS: DIRTY
GIT_COMMIT_DATE: 2023-02-17T13:08:15+01:00
GIT_COMMIT_DESCRIPTION: Add benchmark for Kokkos::for_each
GIT_COMMIT_HASH: 06c406ec2
GPU architecture: none
KOKKOS_COMPILER_CLANG: 1507
KOKKOS_ENABLE_ASM: no
KOKKOS_ENABLE_CXX17: yes
KOKKOS_ENABLE_CXX20: no
KOKKOS_ENABLE_CXX23: no
KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK: yes
KOKKOS_ENABLE_GNU_ATOMICS: no
KOKKOS_ENABLE_HBWSPACE: no
KOKKOS_ENABLE_HWLOC: yes
KOKKOS_ENABLE_INTEL_ATOMICS: no
KOKKOS_ENABLE_INTEL_MM_ALLOC: no
KOKKOS_ENABLE_LIBDL: yes
KOKKOS_ENABLE_LIBRT: no
KOKKOS_ENABLE_PRAGMA_IVDEP: no
KOKKOS_ENABLE_PRAGMA_LOOPCOUNT: no
KOKKOS_ENABLE_PRAGMA_UNROLL: no
KOKKOS_ENABLE_PRAGMA_VECTOR: no
KOKKOS_ENABLE_SERIAL: yes
KOKKOS_ENABLE_SERIAL_ATOMICS: no
KOKKOS_ENABLE_WINDOWS_ATOMICS: no
Kokkos Version: 4.0.99
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations Time (Kokkos) Time (std)
-----------------------------------------------------------------------------------------------------------
Algorithms_ForEach<double>/Size:13           0.000 s         0.000 s       2017120          161n        90n
Algorithms_ForEach<double>/Size:1003         0.000 s         0.000 s        107643        1.995u     4.271u
Algorithms_ForEach<double>/Size:101513       0.001 s         0.001 s          1102      183.739u   427.503u
Algorithms_ForEach<double>/Size:1000003      0.006 s         0.006 s           106      1.98217m   4.35291m
```