@cz4rs 2023-02-17
When using `std::for_each` with `std::execution::par_unseq`:
```
Running ./build/algorithms/unit_tests/KokkosAlgorithms_Benchmark_Kokkos_ForEach
Run on (8 X 4700 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 12288 KiB (x1)
Load Average: 1.14, 1.33, 1.30
CPU architecture: none
Default Device: N6Kokkos6OpenMPE
GIT_BRANCH: benchmark-algorithms
GIT_CLEAN_STATUS: DIRTY
GIT_COMMIT_DATE: 2023-02-17T13:08:15+01:00
GIT_COMMIT_DESCRIPTION: Add benchmark for Kokkos::for_each
GIT_COMMIT_HASH: 06c406ec2
GPU architecture: none
KOKKOS_COMPILER_CLANG: 1507
KOKKOS_ENABLE_ASM: no
KOKKOS_ENABLE_CXX17: yes
KOKKOS_ENABLE_CXX20: no
KOKKOS_ENABLE_CXX23: no
KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK: yes
KOKKOS_ENABLE_GNU_ATOMICS: no
KOKKOS_ENABLE_HBWSPACE: no
KOKKOS_ENABLE_HWLOC: yes
KOKKOS_ENABLE_INTEL_ATOMICS: no
KOKKOS_ENABLE_INTEL_MM_ALLOC: no
KOKKOS_ENABLE_LIBDL: yes
KOKKOS_ENABLE_LIBRT: no
KOKKOS_ENABLE_OPENMP: yes
KOKKOS_ENABLE_OPENMP_ATOMICS: no
KOKKOS_ENABLE_PRAGMA_IVDEP: no
KOKKOS_ENABLE_PRAGMA_LOOPCOUNT: no
KOKKOS_ENABLE_PRAGMA_UNROLL: no
KOKKOS_ENABLE_PRAGMA_VECTOR: no
KOKKOS_ENABLE_WINDOWS_ATOMICS: no
Kokkos: OpenMP thread_pool_topology[ 1 x 8 x 1 ]
Kokkos Version: 4.0.99
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
***WARNING*** Library was built as DEBUG. Timings may be affected.
-----------------------------------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations Time (Kokkos) Time (std)
-----------------------------------------------------------------------------------------------------------
Algorithms_ForEach<double>/Size:13           0.000 s         0.000 s         68595        2.143u       849n
Algorithms_ForEach<double>/Size:1003         0.000 s         0.000 s         11463        8.744u     16.41u
Algorithms_ForEach<double>/Size:101513       0.003 s         0.003 s           243      2.09798m   1.60015m
Algorithms_ForEach<double>/Size:1000003      0.021 s         0.020 s            34      4.14244m  0.0163967
```
Note: these results are with Debug mode enabled. Building in Release fails with:
```
In file included from /home/cz4rs/Kokkos/kokkos/algorithms/unit_tests/BenchmarkStdAlgorithmsForEach.cpp:21:
In file included from /usr/bin/../lib/gcc/x86_64-redhat-linux/12/../../../../include/c++/12/pstl/glue_execution_defs.h:50:
/usr/bin/../lib/gcc/x86_64-redhat-linux/12/../../../../include/c++/12/pstl/algorithm_impl.h:99:1: error: loop not vectorized: the optimizer was unable to perform the requested transformation; the transformation might be disabled or specified as part of an unsupported transformation ordering [-Werror,-Wpass-failed=transform-warning]
```